comments server done

+++ controllers/comment.controllers.js
@@ -71,6 +71,10 @@ const deleteComment = async (req, res) => {
         if (!comment) {
             return res.status(400).json({ msg: "No comment found" })
         }
+        //check if the req.user is the one who creatde the comment
+        if (comment.commentedBy.id.toString() !== req.user._id.toString()) {
+            return res.status(401).json({ msg: "You are not authorized to delete this comment" })
+        }
         await Comment.findByIdAndDelete(commentID);
         res.status(200).json({ msg: "Comment deleted" })
